+++ src/docs/previews/dropdown-menu/main/tailwind/index.svelte
@@ -65,7 +65,7 @@
 		<div class="item" use:melt={$checkboxItem}>
 			<div class="check">
 				{#if $settingsSync}
-					<Check class="icon" />
+					<Check class="square-4" />
 				{/if}
 			</div>
 			Settings Sync is On
@@ -73,7 +73,7 @@
 		<div class="item" use:melt={$subTrigger}>
 			Profiles
 			<div class="rightSlot">
-				<ChevronRight class="icon" />
+				<ChevronRight class="square-4" />
 			</div>
 		</div>
 		{#if $subOpen}
@@ -102,7 +102,7 @@
 		<div class="item" use:melt={$checkboxItemA}>
 			<div class="check">
 				{#if $hideMeltUI}
-					<Check class="icon" />
+					<Check class="square-4" />
 				{/if}
 			</div>
 			Hide Melt UI