+++ src/lib/builders/context-menu/events.ts
@@ -1,4 +1,4 @@
-import type { GroupedEvents } from '$lib/internal/types.js';
+import type { GroupedEvents, MeltComponentEvents } from '$lib/internal/types.js';
 import { menuEvents } from '../menu/events.js';
 
 export const contextMenuEvents = {
@@ -8,3 +8,4 @@ export const contextMenuEvents = {
 };
 
 export type ContextMenuEvents = GroupedEvents<typeof contextMenuEvents>;
+export type ContextMenuComponentEvents = MeltComponentEvents<ContextMenuEvents>;