+++ src/lib/builders/menubar/events.ts
@@ -1,4 +1,4 @@
-import type { GroupedEvents } from '$lib/internal/types.js';
+import type { GroupedEvents, MeltComponentEvents } from '$lib/internal/types.js';
 import { menuEvents } from '../menu/events.js';
 
 export const menubarEvents = {
@@ -8,3 +8,4 @@ export const menubarEvents = {
 };
 
 export type MenubarEvents = GroupedEvents<typeof menubarEvents>;
+export type MenubarComponentEvents = MeltComponentEvents<MenubarEvents>;