+++ src/lib/builders/pagination/events.ts
@@ -1,4 +1,4 @@
-import type { GroupedEvents } from '$lib/internal/types.js';
+import type { GroupedEvents, MeltComponentEvents } from '$lib/internal/types.js';
 
 export const paginationEvents = {
 	pageTrigger: ['click', 'keydown'] as const,
@@ -7,3 +7,4 @@ export const paginationEvents = {
 };
 
 export type PaginationEvents = GroupedEvents<typeof paginationEvents>;
+export type PaginationComponentEvents = MeltComponentEvents<PaginationEvents>;