+++ src/lib/builders/popover/events.ts
@@ -1,4 +1,4 @@
-import type { GroupedEvents } from '$lib/internal/types.js';
+import type { GroupedEvents, MeltComponentEvents } from '$lib/internal/types.js';
 
 export const popoverEvents = {
 	trigger: ['click', 'keydown'] as const,
@@ -6,3 +6,4 @@ export const popoverEvents = {
 };
 
 export type PopoverEvents = GroupedEvents<typeof popoverEvents>;
+export type PopoverComponentEvents = MeltComponentEvents<PopoverEvents>;