+++ src/lib/builders/slider/events.ts
@@ -1,7 +1,8 @@
-import type { GroupedEvents } from '$lib/internal/types.js';
+import type { GroupedEvents, MeltComponentEvents } from '$lib/internal/types.js';
 
 export const sliderEvents = {
 	thumb: ['keydown'] as const,
 };
 
 export type SliderEvents = GroupedEvents<typeof sliderEvents>;
+export type SliderComponentEvents = MeltComponentEvents<SliderEvents>;