+++ src/lib/builders/switch/types.ts
@@ -2,7 +2,7 @@ import type { BuilderReturn } from '$lib/internal/types.js';
 import type { Writable } from 'svelte/store';
 import type { createSwitch } from './create.js';
 import type { ChangeFn } from '$lib/internal/helpers/index.js';
-
+export type { SwitchComponentEvents } from './events.js';
 export type CreateSwitchProps = {
 	/**
 	 * The uncontrolled default checked status of the switch.