+++ src/lib/builders/tags-input/types.ts
@@ -2,7 +2,7 @@ import type { BuilderReturn } from '$lib/internal/types.js';
 import type { Writable } from 'svelte/store';
 import type { createTagsInput } from './create.js';
 import type { ChangeFn } from '$lib/internal/helpers/index.js';
-
+export type { TagsInputComponentEvents } from './events.js';
 export type CreateTagsInputProps = {
 	placeholder?: string;
 	disabled?: boolean;
@@ -37,9 +37,9 @@ export type TagProps = {
 	editable?: boolean;
 };
 
-export type UpdateTag = (tag: Tag) => Promise<Tag>;
-export type RemoveTag = (tag: Tag) => Promise<boolean>;
-export type AddTag = (tag: string) => Promise<Tag | string>;
+export type UpdateTag = (tag: Tag) => Tag | Promise<Tag>;
+export type RemoveTag = (tag: Tag) => boolean | Promise<boolean>;
+export type AddTag = (tag: string) => (Tag | string) | Promise<Tag | string>;
 
 export type TagsInput = BuilderReturn<typeof createTagsInput>;
 export type TagsInputElements = TagsInput['elements'];